requirement.txt
fastapi
uvicorn
PyMuPDF
Pillow
google-generativeai
python-multipart
